fix: resolve all ruff lint and mypy type errors across codebase
Ruff: 3500+ lint issues → 0 (auto-fix + manual fixes + noqa suppression)
Mypy: 193 type errors → 0 (proper typing, None-guards, type-ignore for 3P libs)

Notable real bugs found and fixed:
- news_impact_analyzer.py: operator precedence bug (str - int)
- trading_loop.py: call to nonexistent get_usdrub() method
- bond_cycle.py: added _validated_bond_fields() to guard nullable fields
- telegram_reader.py: None-guard before .replace() on Optional[str]

+++ src/finalayze/backtest/portfolio_orchestrator.py
@@ -365,7 +365,7 @@ class PortfolioBacktestOrchestrator:
         if std <= 0:
             return 0.0
 
-        return mean_excess / std * math.sqrt(_TRADING_DAYS_PER_YEAR)
+        return float(mean_excess / std * math.sqrt(_TRADING_DAYS_PER_YEAR))
 
     @staticmethod
     def _compute_max_drawdown(curve: list[float]) -> float:
